[hotfix][tests] Merge SlotPoolInteractionsTest#TestingSlotPool with TestingSlotPoolImpl
