+++ flink-runtime/src/test/java/org/apache/flink/runtime/jobmaster/slotpool/TestingSlotPoolImpl.java
@@ -22,14 +22,17 @@ import org.apache.flink.api.common.JobID;
 import org.apache.flink.api.common.time.Time;
 import org.apache.flink.configuration.JobManagerOptions;
 import org.apache.flink.runtime.akka.AkkaUtils;
+import org.apache.flink.runtime.clusterframework.types.AllocationID;
 import org.apache.flink.runtime.clusterframework.types.ResourceProfile;
 import org.apache.flink.runtime.jobmaster.SlotRequestId;
+import org.apache.flink.util.Preconditions;
 import org.apache.flink.util.clock.Clock;
 import org.apache.flink.util.clock.SystemClock;
 
 import javax.annotation.Nullable;
 
 import java.util.concurrent.CompletableFuture;
+import java.util.function.Consumer;
 
 /**
  * Testing extension of the {@link SlotPoolImpl} which adds additional methods
@@ -38,6 +41,10 @@ import java.util.concurrent.CompletableFuture;
 public class TestingSlotPoolImpl extends SlotPoolImpl {
 
 	private ResourceProfile lastRequestedSlotResourceProfile;
+
+	private volatile Consumer<SlotRequestId> releaseSlotConsumer;
+
+	private volatile Consumer<SlotRequestId> timeoutPendingSlotRequestConsumer;
 
 	public TestingSlotPoolImpl(JobID jobId) {
 		this(
@@ -54,7 +61,10 @@ public class TestingSlotPoolImpl extends SlotPoolImpl {
 			Time rpcTimeout,
 			Time idleSlotTimeout,
 			Time batchSlotTimeout) {
+
 		super(jobId, clock, rpcTimeout, idleSlotTimeout, batchSlotTimeout);
+		releaseSlotConsumer = null;
+		timeoutPendingSlotRequestConsumer = null;
 	}
 
 	void triggerCheckIdleSlot() {
@@ -82,5 +92,54 @@ public class TestingSlotPoolImpl extends SlotPoolImpl {
 
 	public ResourceProfile getLastRequestedSlotResourceProfile() {
 		return lastRequestedSlotResourceProfile;
+	}
+
+	public void setReleaseSlotConsumer(Consumer<SlotRequestId> releaseSlotConsumer) {
+		this.releaseSlotConsumer = Preconditions.checkNotNull(releaseSlotConsumer);
+	}
+
+	public void setTimeoutPendingSlotRequestConsumer(Consumer<SlotRequestId> timeoutPendingSlotRequestConsumer) {
+		this.timeoutPendingSlotRequestConsumer = Preconditions.checkNotNull(timeoutPendingSlotRequestConsumer);
+	}
+
+	@Override
+	public void releaseSlot(
+			SlotRequestId slotRequestId,
+			@Nullable Throwable cause) {
+
+		final Consumer<SlotRequestId> currentReleaseSlotConsumer = releaseSlotConsumer;
+
+		super.releaseSlot(slotRequestId, cause);
+
+		if (currentReleaseSlotConsumer != null) {
+			currentReleaseSlotConsumer.accept(slotRequestId);
+		}
+	}
+
+	@Override
+	protected void timeoutPendingSlotRequest(SlotRequestId slotRequestId) {
+		final Consumer<SlotRequestId> currentTimeoutPendingSlotRequestConsumer = timeoutPendingSlotRequestConsumer;
+
+		super.timeoutPendingSlotRequest(slotRequestId);
+
+		if (currentTimeoutPendingSlotRequestConsumer != null) {
+			currentTimeoutPendingSlotRequestConsumer.accept(slotRequestId);
+		}
+	}
+
+	boolean containsAllocatedSlot(AllocationID allocationId) {
+		return getAllocatedSlots().contains(allocationId);
+	}
+
+	boolean containsAvailableSlot(AllocationID allocationId) {
+		return getAvailableSlots().contains(allocationId);
+	}
+
+	int getNumberOfPendingRequests() {
+		return getPendingRequests().size();
+	}
+
+	int getNumberOfWaitingForResourceRequests() {
+		return getWaitingForResourceManager().size();
 	}
 }
